refactor: move deterministic git-hook logic out of skills.installer
Fixes a service-layer boundary violation flagged on PR #381 (#352, E4):
services must not import skills, but bootstrap.py imported git-hook
build/install/reconcile logic from wade.skills.installer.

Moves the git-hook cluster verbatim into a new git/hooks.py (git layer),
and the packaged template-asset loaders it depends on into a new leaf
module, utils/templates.py. skills/installer.py keeps thin back-compat
re-exports of the loaders only — the git-hook functions are not
re-exported, severing them from skills for good.

Closes #382
